+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Features
 
+* `[jest-message-util]` Don't ignore messages with `vendor` anymore
+  ([#XXXX](https://github.com/facebook/jest/pull/XXXX))
 * `[jest-validate]` Get rid of `jest-config` dependency
   ([#6067](https://github.com/facebook/jest/pull/6067))
 * `[jest-validate]` Adds option to inject `deprecationEntries`